server/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.4
djangorestframework==3.13.1
filelock==3.7.0
idna==3.3
pipenv==2022.5.2
platformdirs==2.5.2
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
